Moved Cigarette in to a common component
