+++ App/components/index.ts
@@ -19,6 +19,7 @@ export * from './Banner';
 export * from './BoxButton';
 export * from './Button';
 export * from './Cigarettes';
+export * from './CigaretteBlock';
 export * from './ChangeLocation';
 export * from './CurrentLocation';
 export * from './ListItem';
